Code cleanup (vol. 1)
